gomock: moves mock_gomock to an internal directory
mock_gomock is a test artifact. Moving it to an internal directory to
avoid having it exported.
